Check project expire date when files are uploaded.
